Fix text truncation issue in DialogBox
DialogBox was using lavel to display msg, which led to text
being truncated when the message length exceeded the avail
width. This resulted in an inconsistent and frustrating ux
as longer messages were cut off rather than wraped to next
line.

TextFlow provides better control over text layout and
wrapping. Unlike label, it will automatically adjust
the text within the width, so messages are displayed
in full without truncation.

Let's replace Label with TextFlow in DialogBox to:
- Improve text wrapping and prevent truncation
- Allow dynamic resizing and better handling of long msg

The max width for TextFLow is set to 500px to maintain a
clean layout without strething too wide. So the bubbles
are still easy to read.

This approach was chosen because TextFlow offers more
flexibility and control compared to label.